Convert API surface to a .d.ts module for publishing
