+++ src/api/adapters/clusterprovider/apibroker.ts
@@ -1,5 +1,7 @@
-import { APIVersion } from '../../api/api';
-import * as v1 from '../../api/clusterprovider/v1';
+/// <reference path='../../vscode-kubernetes-tools-api.d.ts' />
+
+import { APIVersion } from 'api';
+import * as v1 from 'api.clusterprovider.v1';
 import * as v1adapter from './v1.adapter';
 
 export function api(version: string): APIVersion {
