+++ src/api/adapters/clusterprovider/v1.adapter.ts
@@ -1,5 +1,7 @@
+/// <reference path='../../vscode-kubernetes-tools-api.d.ts' />
+
 import { get as getClusterProviderRegistry } from '../../../components/clusterprovider/clusterproviderregistry';
-import * as v1 from '../../api/clusterprovider/v1';
+import * as v1 from 'api.clusterprovider.v1';
 
 export function implementation(): v1.API {
     return impl;
