--- src/api/api/clusterprovider/component.ts
@@ -1,1 +1,0 @@
-export const componentId = 'clusterprovider';
